Initial Readme
We stored tweets in the database related to our project. We came up with 5 non-trivial questions related to our subject using Twitter data. We implemented data and queries in SQL and created a model diagram representing the relationships and we scripted a bot that scraps Twitter to gather the tweet data for us.